feat: :sparkles: enhance metrics with raw logging, reasoning tokens, and int64 support (014, 015, 020, 022, 023)

+++ internal/model/log.go
@@ -33,8 +33,8 @@ type RelayLog struct {
 	ChannelName      string           `json:"channel_name"`                             // 渠道名称
 	ChannelType      int              `json:"channel_type"`                             // 渠道类型
 	ActualModelName  string           `json:"actual_model_name"`                        // 实际使用模型名称
-	InputTokens      int              `json:"input_tokens"`                             // 输入Token
-	OutputTokens     int              `json:"output_tokens"`                            // 输出 Token
+	InputTokens      int64            `json:"input_tokens"`                             // 输入Token
+	OutputTokens     int64            `json:"output_tokens"`                            // 输出 Token
 	Ftut             int              `json:"ftut"`                                     // 首字时间(毫秒)
 	UseTime          int              `json:"use_time"`                                 // 总用时(毫秒)
 	Cost             float64          `json:"cost"`                                     // 消耗费用
